requirements.txt
streamlit
ultralytics
pillow
onnxruntime